[TMVA] CV - Rename CrossEval tutorials to CrossVal
